requirements.txt
aioredis==2.0.1
aioredis-lock==0.1.0
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.2.0
distlib==0.3.6
Django==4.1.5
django-celery==3.1.17
django-celery-beat==2.4.0
django-timezone-field==5.0
dnspython==2.2.1
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
hiredis==2.1.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
kombu==5.2.4
lxml==4.9.2
mongo==0.2.0
motor==3.1.1
mypy-extensions==0.4.3
packaging==22.0
parsel==1.7.0
pathspec==0.10.3
pipenv==2022.12.19
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymongo==4.3.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7
redis==4.4.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
w3lib==2.1.1
wcwidth==0.2.5
